+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file includes only changes we consider noteworthy for users, admins and plu
 - Fix bug where configured skin logo wasn't loaded via `static.php` resulting in 404 error (#10191)
 - Fix an infinite loop in TNEF (winmail.dat) decoder (#10193)
 - Fix bug where installto.sh would fail if public_html folder does not exist in the target directory (#10202)
+- Revert "Prefer 8bit over quoted-printable for HTML parts, when force_7bit is disabled (#8477)" (#10198)
 
 ## Release 1.7.1